Moved test case to a separate class
